requirements.txt
Flask==2.3.2
transformers==4.48.2
torch==2.0.1
google-api-python-client==2.86.0
google-auth-oauthlib==1.0.0
google-auth-httplib2==0.1.0
twilio==8.3.0
streamlit==1.26.0